Header: wrap AddressBar in an AdwClamp
Currently the addressbar is very small on desktop sizes.
This makes it awkward to reach for and hard to read.
This commit changes the addressbar so that it can
expand up to a maximum size via AdwClamp, while
still being able to fit smaller screens.

We need to set `hhomogeneous` to false on the stacks
in order to have our address bar be nicely centered.